Mummi and bilbo contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl), r00tkiller (Alexandre),  vapaamies (Vladislav Javadov), epryka (Vova - Vladimir), theaifam5 (Mateus), doraemon (currently... also gosu - hiiamgosu, dum - noticemesenpaiiamdum..., litAsHellhahah), mushmouth (Patrick McFarling),  sonnystar (sonnystaminecraft, Vlad Ionescu), drmjg, abbyyyyy, aleksy, alexcode, andrevdb, anita, ayvatan, b3n0it, beee, bigboom, blackburn1911, blackhatisme, boza, camwhite (Cameron White - YOSHICREW | ng2/node + mixing fire, Yoshicrew.com-out-loud contribution...), chrifrachid, code4life, codeotea, cookiesliyr, chubz, cubor, danielv, david_joffe, denvercowboy, diagram, dopedopedope, dzetko, dzhekson6000, euphemisminc, fauker, ffabiorj, georgeb, gillie, gzj71ck4elm2,  hudwan, igsky, intelijens (Michael Bui), ishaan007lucky, jarzi, jilizart, johnyblaze, jookeer, kaeris, karrmarr, kaznowski, kimliksiz, kissloryshy, kranalex, LaffayetteBrittoPhoto (streamer with its C300 photo), magrib, marianyc, matroskinb, michaeljcalkins, michaelphp, milurabbit, monicaa,  nuke, paezrice, peter_cados, phanxgames (Henry Price), purpleizlegitx, rooney2, sanyakozmey, silvaesouza, siox, skycis, skyfoxx, smart6307 (thunder will/canal2048), sova, systemis, tarikogut, thiagovgamg, tonycode, trello, usandy, v0n3x, wanket, wapaca, web2ls, yanaseattle, zaaz and zzzel (Marko Zel)  contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! AlfredoPlays (and his mates and Fwends playing Siege: Above_Perfection, koopz, Macky43 ... ), Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Matty and Dylan), 2016pocket, Bialyorzeu, cary205, cpmgamingltd, esmc103, GameDevCompany (Björn Swedish-Speaking-Finlandian), GrubbyJuicde, Handmadehero (Casey Muratori), kimliksiz, Letorusii, Mattyplays6, michaeljcalkins, nothings2 (Sean Barrett), Shathan, Snipe0803, sundz98 (from Sweden), talus7, Vjzelynxykat, Wheeze202 (and hosting grumbul (British accent?)) and xandergamingtwich contribution!! Beam.pro contribution!! RoxenRox and Xellar contribution!! bisba contribution!! caro  and lucinho contribution!!
Thank you Camilo!

bilbo at veriverse
i7 6700 32GB RAM Windows 10 Pro
(booting also macOS Sierra, Hyper-V and Ubuntu 16.04 Unity Numix Theme Numix Circle Icons)

+++ appseed/aura/aura/os/windows/windows_gdi.cpp
@@ -297,7 +297,7 @@ namespace draw2d
 
          size sSample;
 
-         int maxarea;
+         int64_t maxarea;
 
          ::draw2d::font::e_cs ecs;
